python-backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
opencv-python==4.8.1.78
tensorflow>=2.16.1
pillow>=10.1.0
pillow-heif>=0.13.0
numpy>=1.24.3
mysql-connector-python>=8.2.0
pydantic>=2.5.0
email-validator>=2.1.0